修改应用、添加分享朗读本、添加webview、添加ai批改控制、老师文件分享自动下载

+++ app/src/main/java/com/bll/lnkstudy/utils/FileUtils.java
@@ -269,6 +269,8 @@ public class FileUtils {
      * @param path
      */
     public static void delete(String path){
+        if (path == null || path.isEmpty() )
+            return;
         File file=new File(path);
         deleteFile(file);
     }
